+++ src/map.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 #include <time.h>
 
@@ -14,8 +15,8 @@
 SDL_Rect select_tile[16];
 
 static void MapLoadSelectionTiles(){
-    for(int i=0; i < 4; i++){
-        for(int k=0; k < 4; k++){
+    for(uint32_t i=0; i < 4; i++){
+        for(uint32_t k=0; k < 4; k++){
             // Go row by row in the image
             select_tile[i*4 + k].x = TILE_SIZE * k;
             select_tile[i*4 + k].y = TILE_SIZE * i;
@@ -32,7 +33,7 @@ map_t* MapCreateFromFile(FILE* map_file, SDL_Renderer* renderer){
     map->width = TILE_MAP_MAX_X;
 
     map->tile_data = malloc(map->width * sizeof(int*));
-    for(int i = 0; i < map->width; i++){
+    for(int32_t i = 0; i < map->width; i++){
         map->tile_data[i] = calloc(map->height, sizeof(int));
     }
 
@@ -40,14 +41,14 @@ map_t* MapCreateFromFile(FILE* map_file, SDL_Renderer* renderer){
 
     MapLoadSelectionTiles();
 
-    int row = 0;
-    int max_col = 0;
+    u_int32_t row = 0;
+    u_int32_t max_col = 0;
     char line[1024];
 
     while(fgets(line, sizeof(line), map_file)){
         if(row >= TILE_MAP_MAX_Y) break;
 
-        int col = 0;
+        u_int32_t col = 0;
         char* token = strtok(line, " \n\r;,");
         while(token){
             if(col < TILE_MAP_MAX_X){
@@ -71,20 +72,20 @@ map_t* MapCreateFromFile(FILE* map_file, SDL_Renderer* renderer){
 }
 
 void MapDraw(map_t* map, SDL_Renderer* rend, SDL_Rect viewport){
-    int start_col = viewport.x / 32;
-    int end_col = (viewport.x + viewport.w)/ 32;
-    int start_row = viewport.y / 32;
-    int end_row = (viewport.y + viewport.h) / 32;
+    int32_t start_col = viewport.x / 32;
+    int32_t end_col = (viewport.x + viewport.w)/ 32;
+    int32_t start_row = viewport.y / 32;
+    int32_t end_row = (viewport.y + viewport.h) / 32;
 
     if (start_col < 0) start_col = 0;
     if (start_row < 0) start_row = 0;
     if (end_col >= map->width) end_col = map->width - 1;
     if (end_row >= map->height) end_row = map->height - 1;
 
-    for(int x = start_col; x <= end_col; x++){
-        for(int y = start_row; y <= end_row; y++){
+    for(int32_t x = start_col; x <= end_col; x++){
+        for(int32_t y = start_row; y <= end_row; y++){
             if(x < 0 || x > map->width || y < 0 || y > map->height) continue;
-            int tile_id = map->tile_data[x][y];
+            int8_t tile_id = map->tile_data[x][y];
 
             if(tile_id <= 0 || tile_id > 16) continue;
 
@@ -99,7 +100,7 @@ void MapDraw(map_t* map, SDL_Renderer* rend, SDL_Rect viewport){
     }
 }
 
-void MapUpdateViewport(SDL_Rect* viewport, player_t* player, int map_w_px, int map_h_px, int screen_w, int screen_h){
+void MapUpdateViewport(SDL_Rect* viewport, player_t* player, int32_t map_w_px, int32_t map_h_px, int32_t screen_w, int32_t screen_h){
     viewport->x = player->x_pos - (screen_w / 2);
     viewport->y = player->y_pos - (screen_h / 2);
 
@@ -113,7 +114,7 @@ void MapUpdateViewport(SDL_Rect* viewport, player_t* player, int map_w_px, int m
 void MapDestroy(map_t* map){
     if(!map) return;
 
-    for(int i = 0; i < map->width; i++){
+    for(int32_t i = 0; i < map->width; i++){
         free(map->tile_data[i]);
     }
 
@@ -122,7 +123,7 @@ void MapDestroy(map_t* map){
     free(map);
 }
 
-int GetCurrentTileType(int x_pos, int y_pos, map_t* map){
+int8_t GetCurrentTileType(int32_t x_pos, int32_t y_pos, map_t* map){
     // If somehow someway one of the positions is negative and we want to check the tile type
     // return an error value 
     if(x_pos < 0 || y_pos < 0 || x_pos >= TILE_MAP_MAX_X || y_pos >= TILE_MAP_MAX_Y) return -1;
